Show action button for revenue share correctly
