+++ packages/atlas/src/components/_crt/CloseRevenueShareButton/index.ts
@@ -1,0 +1,1 @@
+export * from './CloseRevenueShareButton'
