+++ packages/atlas/src/components/_inputs/AuctionDatePicker/AuctionDatePicker.tsx
@@ -38,6 +38,7 @@ export type AuctionDatePickerProps = {
   onChange: (value: AuctionDatePickerValue) => void
 } & Omit<SelectProps<AuctionDatePickerValue>, 'onChange'>
 
+// TODO shake animation on date picker is very glitchy, for now just disable it
 export const AuctionDatePicker: FC<AuctionDatePickerProps> = ({
   items,
   value,
